Clean controllers
