+++ app/routes/api.php
@@ -7,17 +7,17 @@
 		Route::get('components/{id}/incidents', 'ComponentController@getComponentIncidents');
 		Route::get('incidents', 'IncidentController@getIncidents');
 		Route::get('incidents/{id}', 'IncidentController@getIncident');
-		Route::get('metrics', 'IncidentController@getMetrics');
-		Route::get('metrics/{id}', 'IncidentController@getMetric');
+		Route::get('metrics', 'MetricController@getMetrics');
+		Route::get('metrics/{id}', 'MetricController@getMetric');
 
 		Route::group(['protected' => true], function() {
 			Route::post('components', 'ComponentController@postComponents');
 			Route::post('incidents', 'IncidentController@postIncidents');
-			Route::post('metrics', 'IncidentController@postMetrics');
+			Route::post('metrics', 'MetricController@postMetrics');
 
 			Route::put('components/{id}', 'ComponentController@putComponent');
 			Route::put('incidents/{id}', 'IncidentController@putIncident');
-			Route::put('metrics/{id}', 'IncidentController@putMetric');
+			Route::put('metrics/{id}', 'MetricController@putMetric');
 		});
 
 	});
